Example Java ACME client with no CN
The maximum length of a CommonName is 64 characters, so technically any
CSR with a longer CN is invalid. Let's Encrypt will reject that today.

But you don't need a CN.  The acme4j CSRBuilder always provides one,
though, so manually constructing a CSR is needed.

This isn't really a complete ACME example - it doesn't have any account
management, asks the user to complete DNS-01 manually, and doesn't save
the key or cert.  But it demonstrates the core request flow.